+++ tests/unit/ed25519.test.ts
@@ -50,13 +50,19 @@ describe("Ed25519PublicKey", () => {
   });
 
   it("should fail malleable signatures", () => {
+    // Here we make a signature exactly with the L
     const signature = new Ed25519Signature(
       // eslint-disable-next-line max-len
-      "ad0828843228048fee9ead35cf0decc8155342e967993851f70bfc0bb9e97719a094c6f956e5cd65872fa769e7dce26f6e637eb4ff70c3f6aafc5f06fbc300ad",
+      "0x0000000000000000000000000000000000000000000000000000000000000000edd3f55c1a631258d69cf7a2def9de1400000000000000000000000000000010",
     );
+    expect(signature.isCanonicalSignature()).toBe(false);
 
-    // Verify with malleable message
-    expect(signature.checkSMalleability()).toBe(false);
+    // We now check with L + 1
+    const signature2 = new Ed25519Signature(
+      // eslint-disable-next-line max-len
+      "0x0000000000000000000000000000000000000000000000000000000000000000edd3f55c1a631258d69cf7a2def9de1400000000000000000000000000000011",
+    );
+    expect(signature2.isCanonicalSignature()).toBe(false);
   });
 
   it("should serialize correctly", () => {